# src/generated/swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
...
		"SimpleResponse": {
			"properties": {
				"success": {
					"type": "boolean"
				},
				"error": {
					"type": "string"
				}
			},
			"required": [
				"success"
			],
			"type": "object",
			"additionalProperties": false
		},
...
		"CommunityUpdateResponse": {
			"properties": {
				"id": {
					"type": "string"
				},
				"title": {
					"type": "string"
				},
				"body": {
					"type": "string"
				},
				"imageUrl": {
					"type": "string",
					"x-nullable": true
				},
				"createdAt": {
					"type": "string"
				},
				"updatedAt": {
					"type": "string"
				}
			},
			"required": [
				"id",
				"title",
				"body",
				"createdAt",
				"updatedAt"
			],
			"type": "object",
			"additionalProperties": false
		},
		"CommunityUpdatesListResponse": {
			"properties": {
				"success": {
					"type": "boolean"
				},
				"data": {
					"items": {
						"$ref": "#/definitions/CommunityUpdateResponse"
					},
					"type": "array"
				},
				"error": {
					"type": "string"
				}
			},
			"required": [
				"success"
			],
			"type": "object",
			"additionalProperties": false
		},
		"CreateCommunityUpdateRequest": {
			"properties": {
				"title": {
					"type": "string"
				},
				"body": {
					"type": "string"
				},
				"imageUrl": {
					"type": "string"
				}
			},
			"required": [
				"title",
				"body"
			],
			"type": "object",
			"additionalProperties": false
		},
		"UpdateCommunityUpdateRequest": {
			"properties": {
				"title": {
					"type": "string"
				},
				"body": {
					"type": "string"
				},
				"imageUrl": {
					"type": "string"
				}
			},
			"type": "object",
			"additionalProperties": false
		},
		"UploadImageResponse": {
			"properties": {
				"success": {
					"type": "boolean"
				},
				"url": {
					"type": "string"
				},
				"error": {
					"type": "string"
				}
			},
			"required": [
				"success"
			],
			"type": "object",
			"additionalProperties": false
		},
		"ChatResponse": {
			"properties": {
				"success": {
					"type": "boolean"
				},
				"message": {
					"type": "string"
				},
				"data": {
					"properties": {
						"usage": {
							"properties": {
								"totalTokens": {
									"type": "number",
									"format": "double"
								},
								"completionTokens": {
									"type": "number",
									"format": "double"
								},
								"promptTokens": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"totalTokens",
								"completionTokens",
								"promptTokens"
							],
							"type": "object"
						},
						"model": {
							"type": "string"
						},
						"response": {
							"type": "string"
						}
					},
					"required": [
						"model",
						"response"
					],
					"type": "object"
				},
				"error": {
					"type": "string"
				}
			},
			"required": [
				"success",
				"message"
			],
			"type": "object",
			"additionalProperties": false
		},
		"ChatMessage": {
			"properties": {
				"role": {
					"type": "string",
					"enum": [
						"user",
						"assistant",
						"system"
					],
					"x-nullable": false
				},
				"content": {
					"type": "string"
				}
			},
			"required": [
				"role",
				"content"
			],
			"type": "object",
			"additionalProperties": false
		},
		"ChatRequest": {
			"properties": {
				"messages": {
					"items": {
						"$ref": "#/definitions/ChatMessage"
					},
					"type": "array"
				}
			},
			"required": [
				"messages"
			],
			"type": "object",
			"additionalProperties": false
		},
		"DiscoveryExtractionData": {
			"properties": {
				"organization": {
					"type": "string"
				},
				"contact_name": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"problem": {
					"type": "string"
				},
				"idea": {
					"type": "string"
				},
				"current_process": {
					"type": "string"
				},
				"target_users": {
					"items": {
						"type": "string"
					},
					"type": "array"
				},
				"desired_outcome": {
					"type": "string"
				},
				"requirements": {
					"items": {
						"type": "string"
					},
					"type": "array"
				},
				"existing_systems": {
					"items": {
						"type": "string"
					},
					"type": "array"
				},
				"constraints": {
					"items": {
						"type": "string"
					},
					"type": "array"
				},
				"timeline": {
					"type": "string"
				},
				"budget": {
					"type": "string"
				},
				"open_questions": {
					"items": {
						"type": "string"
					},
					"type": "array"
				},
				"additional_context": {
					"type": "string"
				},
				"ready_for_submission": {
					"type": "boolean"
				}
			},
			"required": [
				"organization",
				"contact_name",
				"email",
				"problem",
				"idea",
				"current_process",
				"target_users",
				"desired_outcome",
				"requirements",
				"existing_systems",
				"constraints",
				"timeline",
				"budget",
				"open_questions",
				"additional_context",
				"ready_for_submission"
			],
			"type": "object",
			"additionalProperties": false
		},
		"DiscoveryExtractionResponse": {
			"properties": {
				"success": {
					"type": "boolean"
				},
				"data": {
					"$ref": "#/definitions/DiscoveryExtractionData"
				},
				"error": {
					"type": "string"
				}
			},
			"required": [
				"success"
			],
			"type": "object",
			"additionalProperties": false
		},
		"ExtractRequest": {
			"properties": {
				"messages": {
					"items": {
						"$ref": "#/definitions/ChatMessage"
					},
					"type": "array"
				}
			},
			"required": [
				"messages"
			],
			"type": "object",
			"additionalProperties": false
		},
		"StreamRequest": {
			"properties": {
				"messages": {
					"items": {
						"$ref": "#/definitions/ChatMessage"
					},
					"type": "array"
				}
			},
			"required": [
				"messages"
			],
			"type": "object",
			"additionalProperties": false
		},
		"SubmitDiscoveryResponse": {
			"properties": {
				"success": {
					"type": "boolean"
				},
				"message": {
					"type": "string"
				},
				"error": {
					"type": "string"
				}
			},
			"required": [
				"success",
				"message"
			],
			"type": "object",
			"additionalProperties": false
		},
		"SubmitDiscoveryRequest": {
			"properties": {
				"organization": {
					"type": "string"
				},
				"contact_name": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"problem": {
					"type": "string"
				},
				"idea": {
					"type": "string"
				},
				"current_process": {
					"type": "string"
				},
				"target_users": {
					"items": {
						"type": "string"
					},
					"type": "array"
				},
				"desired_outcome": {
					"type": "string"
				},
				"requirements": {
					"items": {
						"type": "string"
					},
					"type": "array"
				},
				"existing_systems": {
					"items": {
						"type": "string"
					},
					"type": "array"
				},
				"constraints": {
					"items": {
						"type": "string"
					},
					"type": "array"
				},
				"timeline": {
					"type": "string"
				},
				"budget": {
					"type": "string"
				},
				"open_questions": {
					"items": {
						"type": "string"
					},
					"type": "array"
				},
				"additional_context": {
					"type": "string"
				}
			},
			"required": [
				"contact_name",
				"email"
			],
			"type": "object",
			"additionalProperties": false
		},
...
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"description": "Prisma, TypeScript, Swagger & tsoa Backend",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/community/updates": {
			"get": {
				"operationId": "List",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Updates retrieved successfully",
						"schema": {
							"$ref": "#/definitions/CommunityUpdatesListResponse"
						}
					}
				},
				"description": "Public — list community updates for the site's community page, newest first.",
				"tags": [
					"Community"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Update created successfully",
						"schema": {
							"properties": {
								"error": {
									"type": "string"
								},
								"data": {
									"$ref": "#/definitions/CommunityUpdateResponse"
								},
								"success": {
									"type": "boolean"
								}
							},
							"required": [
								"success"
							],
							"type": "object"
						}
					},
					"400": {
						"description": "Missing required fields"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"consumes": [
					"application/json"
				],
				"description": "Create a community update (admin only).",
				"tags": [
					"Community"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/CreateCommunityUpdateRequest"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/api/community/updates/{id}": {
			"patch": {
				"operationId": "Edit",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Update edited successfully",
						"schema": {
							"properties": {
								"error": {
									"type": "string"
								},
								"data": {
									"$ref": "#/definitions/CommunityUpdateResponse"
								},
								"success": {
									"type": "boolean"
								}
							},
							"required": [
								"success"
							],
							"type": "object"
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"consumes": [
					"application/json"
				],
				"description": "Edit a community update (admin only).",
				"tags": [
					"Community"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/UpdateCommunityUpdateRequest"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"type": "string"
					}
				]
			},
			"delete": {
				"operationId": "Remove",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Update deleted successfully",
						"schema": {
							"$ref": "#/definitions/SimpleResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"description": "Delete a community update (admin only).",
				"tags": [
					"Community"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/api/community/upload-image": {
			"post": {
				"operationId": "UploadImage",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Image uploaded successfully",
						"schema": {
							"$ref": "#/definitions/UploadImageResponse"
						}
					},
					"400": {
						"description": "No file provided"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"consumes": [
					"multipart/form-data"
				],
				"description": "Upload an image to Cloudinary (admin only). Returns the hosted URL to\nuse as a community update's imageUrl.",
				"tags": [
					"Community"
				],
				"security": [],
				"parameters": [
					{
						"in": "formData",
						"name": "image",
						"required": true,
						"type": "file"
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/api/ai/chat": {
			"post": {
				"operationId": "Chat",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "AI response generated successfully",
						"schema": {
							"$ref": "#/definitions/ChatResponse"
						}
					},
					"400": {
						"description": "Bad request - missing or invalid messages"
					},
					"500": {
						"description": "Internal server error"
					}
				},
				"consumes": [
					"application/json"
				],
				"description": "Get AI response from Groq (non-streaming)",
				"tags": [
					"AI"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "request",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ChatRequest"
						}
					}
				]
			}
		},
		"/api/ai/extract": {
			"post": {
				"operationId": "Extract",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Discovery data extracted successfully",
						"schema": {
							"$ref": "#/definitions/DiscoveryExtractionResponse"
						}
					},
					"400": {
						"description": "Bad request - missing or invalid messages"
					},
					"500": {
						"description": "Internal server error"
					}
				},
				"consumes": [
					"application/json"
				],
				"description": "Extract structured discovery data from the full conversation so far.\nThis runs as its own Groq call (JSON mode, separate from Souly's\npersona) so the frontend never has to parse fields out of Souly's\nvisible reply — which by design never contains the raw JSON report.\nThe `ready_for_submission` flag is the frontend's signal to call\n/api/ai/submit; it only becomes true once the user has explicitly\nconfirmed a summary Souly presented back to them.",
				"tags": [
					"AI"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "request",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ExtractRequest"
						}
					}
				]
			}
		},
		"/api/ai/chat/stream": {
			"post": {
				"operationId": "ChatStream",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "",
						"schema": {
							"additionalProperties": true
						}
					},
					"400": {
						"description": "",
						"schema": {
							"additionalProperties": true
						}
					},
					"500": {
						"description": "",
						"schema": {
							"additionalProperties": true
						}
					}
				},
				"consumes": [
					"application/json"
				],
				"description": "Get AI response with streaming support (Server-Sent Events)",
				"tags": [
					"AI"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "request",
						"required": true,
						"schema": {
							"$ref": "#/definitions/StreamRequest"
						}
					}
				]
			}
		},
		"/api/ai/submit": {
			"post": {
				"operationId": "SubmitDiscovery",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Discovery submitted successfully",
						"schema": {
							"$ref": "#/definitions/SubmitDiscoveryResponse"
						}
					},
					"400": {
						"description": "Bad request - missing name or email"
					},
					"422": {
						"description": "Unprocessable entity - invalid data format"
					},
					"500": {
						"description": "Internal server error"
					}
				},
				"consumes": [
					"application/json"
				],
				"description": "Called once the frontend has a confirmed discovery report. Creates a\ncontact record with source=ai_chat so it shows up in the same dashboard\nas regular contact-form submissions, tagged distinctly.",
				"tags": [
					"AI"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "request",
						"required": true,
						"schema": {
							"$ref": "#/definitions/SubmitDiscoveryRequest"
						}
					}
				]
			}
		},
...
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"jwt": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		}
	}
}